pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discos-client"
version = "0.1.0"
description = "Python client for DISCOS"
authors = [{ name = "Giuseppe Carboni", email = "giuseppe.carboni@inaf.it" }]
license = { text = "GPL-3.0-only" }
requires-python = ">=3.9"
readme = "README.rst"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyzmq",
]

[tool.setuptools]
packages = ["discos_client"]

[tool.setuptools.package-data]
"discos_client" = ["schemas/**"]

[project.optional-dependencies]
test = ["coverage", "prospector"]
docs = ["sphinx", "sphinx-rtd-theme", "sphinx-autodoc-typehints", "sphinx-jsonschema"]
